GCS-1 refactor classes
