+++ trainer-hours-service/src/main/java/com/gym/crm/microservices/trainer/hours/service/repository/TrainerSummaryRepository.java
@@ -3,6 +3,7 @@ package com.gym.crm.microservices.trainer.hours.service.repository;
 import com.gym.crm.microservices.trainer.hours.service.entity.TrainerSummary;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
@@ -13,12 +14,15 @@ public interface TrainerSummaryRepository extends JpaRepository<TrainerSummary, 
     Optional<TrainerSummary> findByUsername(String username);
 
     @Query("""
-            SELECT ms.totalTrainingDuration FROM TrainerSummary ts
+            SELECT ms.totalTrainingDuration
+            FROM TrainerSummary ts
             JOIN ts.yearlySummaries ys
             JOIN ys.monthlySummaries ms
-            WHERE ts.username = ?1
-            AND ys.year = ?2
-            AND ms.month = ?3
+            WHERE ts.username = :username
+            AND ys.year = :year
+            AND ms.month = :month
             """)
-    Integer findWorkloadByParams(String username, Integer year, Integer month);
+    Integer findWorkloadByParams(@Param("username") String username,
+                                 @Param("year") Integer year,
+                                 @Param("month") Integer month);
 }
