+++ trainer-hours-service/src/main/java/com/gym/crm/microservices/trainer/hours/service/rest/exception/GlobalExceptionHandler.java
@@ -1,5 +1,6 @@
 package com.gym.crm.microservices.trainer.hours.service.rest.exception;
 
+import com.gym.crm.microservices.trainer.hours.service.exception.DataNotFoundException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
@@ -9,7 +10,7 @@ import static org.springframework.http.HttpStatus.NOT_FOUND;
 @RestControllerAdvice
 public class GlobalExceptionHandler {
 
-    private static final int NOT_FOUND_CODE = 404;
+    private static final int NOT_FOUND_CODE = 20534;
 
     @ExceptionHandler(DataNotFoundException.class)
     public ResponseEntity<ErrorResponse> handleException(DataNotFoundException e) {
