+++ trainer-hours-service/src/main/java/com/gym/crm/microservices/trainer/hours/service/service/TrainerSummaryService.java
@@ -1,7 +1,7 @@
 package com.gym.crm.microservices.trainer.hours.service.service;
 
-import com.gym.crm.microservices.trainer.hours.service.rest.model.TrainerSummaryRequest;
-import com.gym.crm.microservices.trainer.hours.service.rest.model.TrainerWorkloadResponse;
+import com.gym.crm.microservices.trainer.hours.service.model.TrainerSummaryRequest;
+import com.gym.crm.microservices.trainer.hours.service.model.TrainerWorkloadResponse;
 
 public interface TrainerSummaryService {
