+++ trainer-hours-service/src/test/java/com/gym/crm/microservices/trainer/hours/service/service/impl/TrainerSummaryServiceImplTest.java
@@ -4,9 +4,9 @@ import com.gym.crm.microservices.trainer.hours.service.entity.MonthlySummary;
 import com.gym.crm.microservices.trainer.hours.service.entity.TrainerSummary;
 import com.gym.crm.microservices.trainer.hours.service.entity.YearlySummary;
 import com.gym.crm.microservices.trainer.hours.service.repository.TrainerSummaryRepository;
-import com.gym.crm.microservices.trainer.hours.service.rest.exception.DataNotFoundException;
-import com.gym.crm.microservices.trainer.hours.service.rest.model.TrainerSummaryRequest;
-import com.gym.crm.microservices.trainer.hours.service.rest.model.TrainerWorkloadResponse;
+import com.gym.crm.microservices.trainer.hours.service.exception.DataNotFoundException;
+import com.gym.crm.microservices.trainer.hours.service.model.TrainerSummaryRequest;
+import com.gym.crm.microservices.trainer.hours.service.model.TrainerWorkloadResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
